requirements.txt
aiogram==3.0.0b7
python-dotenv==1.0.0
aiosqlite==0.19.0
aiohttp==3.9.1
